Move ResponseParser to anthropic package and update references
